Refactor(performance): Split the cache module
cache.py held 834 lines, past the 400-line reviewability limit the
aislop code-quality gate applies.

Split it into a package at the same dotted path: models (the cache
type enum, entry and stats records), keys (the key-derivation
helpers), manager (the CacheManager storage engine) and typed_caches
(the per-domain facades and the factory). The package __init__
re-exports the previous names, so performance/__init__.py and every
other import site are unchanged.
